rename graph page
